Update fileoperations.h
The header now asks for the filename of the for creating a new one as well as for editing an existing file in the directory(basefilepath).

Also, added two new private variables. One for the 'filename', and one for 'finalpath': a concatenation of 'basefilepath' & 'filename'.

Finally, added 3 private functions: file_empty(), filepath_empty() and convert_finalpath().